+++ packages/payload/package.json
@@ -3,7 +3,7 @@
   "version": "3.0.0-alpha.49",
   "description": "Node, React and MongoDB Headless CMS and Application Framework",
   "license": "MIT",
-  "main": "./src/index.js",
+  "main": "./src/index.ts",
   "types": "./src/index.ts",
   "type": "module",
   "bin": {
