+++ packages/payload/src/translations/getLocalI18n.ts
@@ -1,9 +1,9 @@
 import { initI18n } from '@payloadcms/translations'
 import { translations } from '@payloadcms/translations/api'
 
-import type { SanitizedConfig } from '../exports/types.js'
+import type { SanitizedConfig } from '../config/types.js'
 
-export const getLocalI18n = async ({
+export const getLocalI18n = ({
   config,
   language = 'en',
 }: {
